refactor: SRP-split index and index_produce modules under 50-line limit
- index.rs (243 lines) → index/{mod,types,cache,cache_mut,cache_async,tests,tests_async,tests_summary_for}.rs
- index_produce.rs (64 lines) → index_produce/{mod,build_context,call}.rs
- All 11 new files under 50 effective lines (max: tests.rs at 49)
- All 10 tests pass, re-export paths unchanged
- Fields tree/generation made pub(super) for cross-file impl blocks
- Fixed unused mut warning in empty_index_returns_none_on_get
